feat: remove App.css and restructure components for Aatrox champion view, including new Layout, Navbar, and HeroPage components
