+++ frontend/src/components/Layout.tsx
@@ -7,15 +7,17 @@ interface LayoutProps {
   children: React.ReactNode;
   onShowAatrox: () => void;
   onShowGrid: () => void;
+  onShowHero: () => void;
   currentView: AppView;
 }
 
-const Layout: React.FC<LayoutProps> = ({ children, onShowAatrox, onShowGrid, currentView }) => {
+const Layout: React.FC<LayoutProps> = ({ children, onShowAatrox, onShowGrid, onShowHero, currentView }) => {
   return (
     <div className="min-h-screen bg-neutral-50 dark:bg-neutral-900 text-neutral-900 dark:text-neutral-100">
       <Navbar 
         onShowAatrox={onShowAatrox}
         onShowGrid={onShowGrid}
+        onShowHero={onShowHero}
         currentView={currentView}
       />
       <main className="flex-1">
